add display opponent videoTrack in Audio mode
