+++ sample-videochat-conference/src/main/java/com/quickblox/sample/conference/fragments/AudioConversationFragment.java
@@ -13,6 +13,7 @@ import com.quickblox.sample.conference.R;
 import com.quickblox.sample.conference.activities.CallActivity;
 import com.quickblox.sample.conference.adapters.OpponentsFromCallAdapter;
 import com.quickblox.videochat.webrtc.QBRTCAudioTrack;
+import com.quickblox.videochat.webrtc.callbacks.QBRTCClientAudioTracksCallback;
 
 import java.io.Serializable;
 
@@ -21,7 +22,7 @@ import java.io.Serializable;
  */
 
 public class AudioConversationFragment extends BaseConversationFragment implements Serializable, OpponentsFromCallAdapter.OnAdapterEventListener,
-        CallActivity.OnChangeDynamicToggle {
+        QBRTCClientAudioTracksCallback<ConferenceSession>, CallActivity.OnChangeDynamicToggle {
     private String TAG = getClass().getSimpleName();
 
     private TextView localName;
@@ -95,6 +96,29 @@ public class AudioConversationFragment extends BaseConversationFragment implemen
             audioSwitchToggleButton.setChecked(true);
         } else {
             audioSwitchToggleButton.setChecked(false);
+        }
+    }
+
+    @Override
+    protected void initTrackListeners() {
+        super.initTrackListeners();
+        initAudioTracksListener();
+    }
+
+    @Override
+    protected void removeTrackListeners() {
+        removeAudioTracksListener();
+    }
+
+    private void initAudioTracksListener() {
+        if (currentSession != null) {
+            currentSession.addAudioTrackCallbacksListener(this);
+        }
+    }
+
+    private void removeAudioTracksListener() {
+        if (currentSession != null) {
+            currentSession.removeAudioTrackCallbacksListener(this);
         }
     }
 }
