+++ sample-videochat-conference/src/main/java/com/quickblox/sample/conference/fragments/VideoConversationFragment.java
@@ -12,24 +12,18 @@ import com.quickblox.conference.view.QBConferenceSurfaceView;
 import com.quickblox.sample.conference.R;
 import com.quickblox.sample.conference.adapters.OpponentsFromCallAdapter;
 import com.quickblox.sample.core.utils.Toaster;
-import com.quickblox.videochat.webrtc.QBRTCAudioTrack;
-import com.quickblox.videochat.webrtc.callbacks.QBRTCClientVideoTracksCallbacks;
 import com.quickblox.videochat.webrtc.callbacks.QBRTCSessionStateCallback;
 import com.quickblox.videochat.webrtc.view.QBRTCVideoTrack;
 
 import org.webrtc.CameraVideoCapturer;
-import org.webrtc.RendererCommon;
-import org.webrtc.SurfaceViewRenderer;
 import org.webrtc.VideoRenderer;
 
 import java.io.Serializable;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * QuickBlox team
  */
-public class VideoConversationFragment extends BaseConversationFragment implements Serializable, QBRTCClientVideoTracksCallbacks<ConferenceSession>,
+public class VideoConversationFragment extends BaseConversationFragment implements Serializable,
         QBRTCSessionStateCallback<ConferenceSession>, OpponentsFromCallAdapter.OnAdapterEventListener {
 
     private String TAG = VideoConversationFragment.class.getSimpleName();
@@ -38,8 +32,7 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
     private CameraState cameraState = CameraState.DISABLED_FROM_USER;
 
     private QBRTCVideoTrack localVideoTrack;
-    private Map<Integer, QBRTCVideoTrack> videoTrackMap;
-    private boolean isCurrentCameraFront;
+    protected boolean isCurrentCameraFront;
 
 
     @Override
@@ -62,18 +55,6 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
     public void onStart() {
         super.onStart();
         Log.i(TAG, "onStart");
-    }
-
-    @Override
-    protected void initTrackListeners() {
-        super.initTrackListeners();
-        initVideoTracksListener();
-    }
-
-    @Override
-    protected void removeTrackListeners() {
-        super.removeTrackListeners();
-        removeVideoTracksListener();
     }
 
     @Override
@@ -121,9 +102,6 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
     public void onDestroyView() {
         super.onDestroyView();
         Log.d(TAG, "onDestroyView");
-        removeVideoTracksListener();
-        removeVideoTrackRenderers();
-        releaseViews();
     }
 
     @Override
@@ -132,17 +110,6 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
         cameraToggle.setVisibility(View.INVISIBLE);
     }
 
-    private void initVideoTracksListener() {
-        if (currentSession != null) {
-            currentSession.addVideoTrackCallbacksListener(this);
-        }
-    }
-
-    private void initAudioTracksListener() {
-        if (currentSession != null) {
-            currentSession.addAudioTrackCallbacksListener(this);
-        }
-    }
 
     protected void initButtonsListener() {
         super.initButtonsListener();
@@ -205,6 +172,15 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
         }
     }
 
+    @Override
+    protected void fillVideoView(QBConferenceSurfaceView videoView, QBRTCVideoTrack videoTrack,
+                                 boolean remoteRenderer) {
+        super.fillVideoView(videoView, videoTrack, remoteRenderer);
+        if (!remoteRenderer) {
+            updateVideoView(videoView, isCurrentCameraFront);
+        }
+    }
+
     ////////////////////////////  callbacks from QBRTCClientVideoTracksCallbacks ///////////////////
     @Override
     public void onLocalVideoTrackReceive(ConferenceSession qbrtcSession, final QBRTCVideoTrack videoTrack) {
@@ -218,106 +194,7 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
         }
     }
 
-    @Override
-    public void onRemoteVideoTrackReceive(ConferenceSession session, final QBRTCVideoTrack videoTrack, final Integer userID) {
-        Log.d(TAG, "onRemoteVideoTrackReceive for opponent= " + userID);
-        getVideoTrackMap().put(userID, videoTrack);
-    }
-
-
-    @Override
-    public void onLocalAudioTrackReceive(ConferenceSession session, QBRTCAudioTrack audioTrack) {
-    }
-
-    @Override
-    public void onRemoteAudioTrackReceive(ConferenceSession session, QBRTCAudioTrack audioTrack, Integer userID) {
-    }
-
     /////////////////////////////////////////    end    ////////////////////////////////////////////
-
-    @Override
-    protected void setRemoteViewMultiCall(int userID) {
-        super.setRemoteViewMultiCall(userID);
-        Log.d(TAG, "setRemoteViewMultiCall fillVideoView");
-
-        final OpponentsFromCallAdapter.ViewHolder itemHolder = getViewHolderForOpponent(userID);
-        if (itemHolder == null) {
-            Log.d(TAG, "itemHolder == null - true");
-            return;
-        }
-        final QBConferenceSurfaceView remoteVideoView = itemHolder.getOpponentView();
-
-        if (remoteVideoView != null) {
-            remoteVideoView.setZOrderMediaOverlay(true);
-            updateVideoView(remoteVideoView, false);
-            Log.d(TAG, "onRemoteVideoTrackReceive fillVideoView");
-            QBRTCVideoTrack remoteVideoTrack = getVideoTrackMap().get(userID);
-            if(remoteVideoTrack != null){
-                fillVideoView(remoteVideoView, remoteVideoTrack, true);
-            }
-        }
-    }
-
-    private void fillVideoView(QBConferenceSurfaceView videoView, QBRTCVideoTrack videoTrack,
-                               boolean remoteRenderer) {
-        videoTrack.removeRenderer(videoTrack.getRenderer());
-        videoTrack.addRenderer(new VideoRenderer(videoView));
-        if (!remoteRenderer) {
-            updateVideoView(videoView, isCurrentCameraFront);
-        }
-        Log.d(TAG, (remoteRenderer ? "remote" : "local") + " Track is rendering");
-    }
-
-    protected void updateVideoView(SurfaceViewRenderer surfaceViewRenderer, boolean mirror) {
-        updateVideoView(surfaceViewRenderer, mirror, RendererCommon.ScalingType.SCALE_ASPECT_FILL);
-    }
-
-    protected void updateVideoView(SurfaceViewRenderer surfaceViewRenderer, boolean mirror, RendererCommon.ScalingType scalingType) {
-        Log.i(TAG, "updateVideoView mirror:" + mirror + ", scalingType = " + scalingType);
-        surfaceViewRenderer.setScalingType(scalingType);
-        surfaceViewRenderer.setMirror(mirror);
-        surfaceViewRenderer.requestLayout();
-    }
-
-    protected Map<Integer, QBRTCVideoTrack> getVideoTrackMap() {
-        if (videoTrackMap == null) {
-            videoTrackMap = new HashMap<>();
-        }
-        return videoTrackMap;
-    }
-
-    private void removeVideoTrackRenderers() {
-        Log.d(TAG, "removeVideoTrackRenderers");
-        Log.d(TAG, "remove opponents video Tracks");
-        Map<Integer, QBRTCVideoTrack> videoTrackMap = getVideoTrackMap();
-        for (QBRTCVideoTrack videoTrack : videoTrackMap.values()) {
-            if (videoTrack.getRenderer() != null) {
-                Log.d(TAG, "remove opponent video Tracks");
-                videoTrack.removeRenderer(videoTrack.getRenderer());
-            }
-        }
-    }
-
-    private void releaseViews() {
-        if (localVideoView != null) {
-            localVideoView.release();
-        }
-        localVideoView = null;
-
-        releaseOpponentsViews();
-    }
-
-    private void removeVideoTracksListener() {
-        if (currentSession != null) {
-            currentSession.removeVideoTrackCallbacksListener(this);
-        }
-    }
-
-    @Override
-    protected void cleanUpAdapter(int userId){
-        super.cleanUpAdapter(userId);
-        getVideoTrackMap().remove(userId);
-    }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -337,5 +214,3 @@ public class VideoConversationFragment extends BaseConversationFragment implemen
         ENABLED_FROM_USER
     }
 }
-
-
