+++ sample-videochat-conference/src/main/res/menu/activity_selected_opponents.xml
@@ -19,7 +19,7 @@
 
     <item
         android:id="@+id/start_as_listener"
-        android:title="@string/action_start_audio_call"
+        android:title="@string/action_start_broadcast_call"
         android:icon="@drawable/ic_ear_hearing_grey600_24dp"
         app:showAsAction="always"
         tools:ignore="AlwaysShowAction" />
